Added 2 methods
can reformat index and recursively add directories as trees

+++ Git.java
@@ -1,7 +1,9 @@
 import java.io.BufferedInputStream;
+import java.io.BufferedReader;
 import java.io.BufferedWriter;
 import java.io.File;
 import java.io.FileInputStream;
+import java.io.FileReader;
 import java.io.IOException;
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
@@ -9,6 +11,7 @@ import java.io.FileWriter;
 import java.math.BigInteger;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import java.nio.file.Paths;
 
 public class Git {
     public static void main(String[] args) throws IOException {
@@ -19,11 +22,33 @@ public class Git {
         Git git = new Git();
 
         // tests computeSHA1 method
-        System.out.println(computeSHA1("test.txt"));
+        File testFile = new File("test.txt");
+        //System.out.println(computeSHA1("test.txt"));
 
-        git.add("test.txt");
-        git.add("test2.txt");
+        testFile = new File("tester/1.txt");
+        //System.out.println(computeSHA1(testFile.getAbsolutePath()));
+        //System.out.println(computeSHA1("tester"));
+        // System.out.println(computeSHA1("tester/empty") + "\n");
+
+        // files u created
+        // git.add("test.txt");
+        // git.add("test2.txt");
+
+        // files i made in tester folder
+        // System.out.println("Testing Git.add(String)...");
+        // System.out.println("");
+        // git.add("tester/1.txt");
+        // git.add("tester/2.txt");
+        // git.add("tester/wahoo/coolFile.txt");
+        // git.add("tester/emptyFile.txt");
+        // git.add("");
+
+        System.out.println("\naddDirectory(String) test... Should print hash: " + git.addDirectory("tester"));
+        git.reformatIndex();
     }
+
+    // if the printed messages are too wordy/redundant, u can set printMessages to false (or just delete them)
+    public static boolean printMessages = true;
 
     // Git constructor checks if a git folder already exists in the parent directory
     // if a git directory doesn't exist, :
@@ -49,15 +74,19 @@ public class Git {
     }
 
     public void add(String input) throws IOException {
-        String fileHash = computeSHA1(input);
-        File hashFile = new File("git/objects/" + fileHash);
-        if (!hashFile.exists()) {
-            Files.copy(Path.of(input), Path.of(hashFile.getPath()));
+        if (printMessages && !new File(input).exists()) { // just for testing; u can delete later if u don't want it
+            System.out.println("Input file does not exist.");
+        } else {
+            String fileHash = computeSHA1(input);
+            File hashFile = new File("git/objects/" + fileHash);
+            if (!hashFile.exists()) {
+                Files.copy(Path.of(input), Path.of(hashFile.getPath()));
+            }
+            BufferedWriter writer = new BufferedWriter(new FileWriter("git/index", true));
+            writer.write(fileHash + " " + input);
+            writer.newLine();
+            writer.close();
         }
-        BufferedWriter writer = new BufferedWriter(new FileWriter("git/index", true));
-        writer.write(fileHash + " " + input);
-        writer.newLine();
-        writer.close();
     }
 
     // deletes git folder and all files inside
@@ -88,6 +117,10 @@ public class Git {
     // copied from geek for geek
     // https://www.geeksforgeeks.org/sha-1-hash-in-java/
     public static String computeSHA1(String input) throws IOException {
+        File inputFile = new File(input);
+        if (printMessages && !new File(input).exists()) { // also just for testing, SORRY
+            return "Cannot compute Sha1; Input file does not exist.";
+        }
         try {
             // getInstance() method is called with algorithm SHA-1
             MessageDigest md = MessageDigest.getInstance("SHA-1");
@@ -96,7 +129,6 @@ public class Git {
             // to calculate message digest of the input string
             // returned as array of byte
             BufferedInputStream stream = new BufferedInputStream(new FileInputStream(input));
-            File inputFile = new File(input);
             byte[] arr = new byte[(int) inputFile.length()];
             stream.read(arr);
             stream.close();
@@ -121,5 +153,93 @@ public class Git {
         catch (NoSuchAlgorithmException e) {
             throw new RuntimeException(e);
         }
+    }
+
+    // reformats the index txt file by replacing the old index with an updated version.
+    public boolean reformatIndex() {
+        try {
+            File treeFormatIndex = new File("git/newIndex");
+            BufferedReader reader = new BufferedReader(new FileReader("git/index"));
+            BufferedWriter writer = new BufferedWriter(new FileWriter("git/newIndex"));
+
+            String currentLine = reader.readLine();
+            while (currentLine != null) {
+                File currentLineFile = new File(currentLine.substring(41));
+                if (!currentLine.substring(0,4).equals("tree") && !currentLine.substring(0,4).equals("blob")) {
+                    if (currentLineFile.exists() && currentLineFile.isFile()) {
+                        writer.write("blob ");
+                    }
+                }
+                writer.write(currentLine);
+                writer.newLine();
+                currentLine = reader.readLine();
+            }
+
+            File oldIndex = new File("git/index");
+            oldIndex.delete();
+
+            treeFormatIndex.createNewFile();
+            treeFormatIndex.renameTo(oldIndex);
+            
+            reader.close();
+            writer.close();
+            return true;
+
+        } catch (IOException e) {
+            e.printStackTrace();
+            return false;
+        }
+        
+    }
+
+    public String addDirectory(String directoryPath) throws IOException {
+        if (!new File(directoryPath).canWrite()) {
+            System.out.println("The given directory is inaccessible.");
+            return "";
+        }
+        if (!new File(directoryPath).exists()) {
+            System.out.println("The given directory does not exist.");
+            return "";
+        }
+        String hashOfTree = recursiveAddDirectory(directoryPath, 0);
+
+        if (printMessages) System.out.println("Successfully added directory!");
+        return hashOfTree;
+    }
+
+    // recursively adds given directory while also adding its own files and subdirectories.
+    // the numberOfTrees parameter helps keep track of which tree's file you are currently editing.
+    private String recursiveAddDirectory(String directoryPath, int numberOfTrees) throws IOException {
+        File currentDirectory = new File(directoryPath);
+        File[] listOfContents = currentDirectory.listFiles();
+
+        if (listOfContents == null) {
+            add(directoryPath);
+            return Git.computeSHA1(directoryPath);
+        }
+
+        File treeFile = new File("git/objects/tree" + numberOfTrees);
+        BufferedWriter writeToTree = new BufferedWriter(new FileWriter(treeFile));
+        
+        for (File currFile : listOfContents) {
+            if (currFile.isFile()) {
+                writeToTree.write("blob " + recursiveAddDirectory(currFile.getPath(), numberOfTrees+1) + " " + currFile.getName());
+                writeToTree.newLine();
+            } else {
+                writeToTree.write("tree " + recursiveAddDirectory(currFile.getPath(), numberOfTrees+1) + " " + currFile.getName());
+                writeToTree.newLine();
+            }
+        }
+        writeToTree.close();
+
+        String hashOfTree = computeSHA1("git/objects/tree"+numberOfTrees);
+        treeFile.renameTo(new File("git/objects/" + computeSHA1("git/objects/tree" + numberOfTrees)));
+
+        BufferedWriter writer = new BufferedWriter(new FileWriter("git/index", true));
+        writer.write("tree " + hashOfTree + " " + directoryPath);
+        writer.newLine();
+        writer.close();
+
+        return hashOfTree;
     }
 }
